query-engine/master/requirements.txt
grpcio
grpcio-tools
sqlparse